requirements.txt
aiobotocore==2.12.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
botocore==1.34.51
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dataclasses-json==0.6.4
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
fsspec==2023.12.2
gcsfs==2023.12.2.post1
google-api-core==2.17.1
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.9
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.7
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.1
langsmith==0.1.27
lxml==5.1.0
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.14.1
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
pinecone-client==3.1.0
pinecone-datasets==0.7.0
protobuf==4.25.3
py4j==0.10.9.5
pyarrow==11.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.3
pypdfium2==4.28.0
pyspark==3.3.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
s3fs==2023.12.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
